feat: + DRY <GameToolkitButton/> component
- Refactor WET Game Toolkits template to use new, reusable DRY component
- Component emits 'update:totalSearchResults' event to render DB search results:
    - watcher based on computed 'displayedToolkits' length
    - Added props for toolkits, searchQuery, showGallery (flag), and totalSearchResults
