+++ src/pages/GameToolkit/GameToolkitButton.vue
@@ -1,0 +1,133 @@
+<script setup>
+/* GameToolkitButton.vue
+ *
+ * - Reusable UI button to display and download selected Game Toolkits
+ * - Component used in Gallery and Database sections
+ */
+
+import { computed, defineEmits, watch } from 'vue';
+
+/* --- PROPS --- */
+
+const props = defineProps({
+  toolkits: {
+    type: Array, // Array of dictionaries per Game Toolkit
+    default: () => [],
+  },
+  searchQuery: {
+    type: String,
+    default: '',
+  },
+  showGallery: {
+    type: Boolean,
+    default: true, // If 'false': Show search results from toolkits DB
+  },
+  totalSearchResults: {
+    type: Number,
+    default: 8,
+  },
+});
+
+const displayedToolkits = computed(() => {
+  if (props.showGallery || !props.searchQuery) {
+    return props.toolkits.slice(0, 8);
+  }
+  return props.toolkits.filter(
+    (toolkit) =>
+      toolkit.name.toLowerCase().includes(props.searchQuery.toLowerCase()) ||
+      toolkit.description
+        .toLowerCase()
+        .includes(props.searchQuery.toLowerCase()) ||
+      toolkit.category.toLowerCase().includes(props.searchQuery.toLowerCase())
+  );
+});
+
+// Emit event to render total search results in Toolkits DB search bar
+const emit = defineEmits(['update:totalSearchResults']);
+
+watch(displayedToolkits, (newList) => {
+  emit('update:totalSearchResults', newList.length);
+});
+
+const openToolkitPDF = (pdfUrl, gameName) => {
+  if (pdfUrl) {
+    window.open(pdfUrl, '_blank');
+  } else {
+    console.warn(`No PDF available for ${gameName}`);
+    alert(`PDF not available for ${gameName} yet. Coming soon!`);
+  }
+};
+</script>
+
+<template>
+  <button
+    v-for="toolkit in displayedToolkits"
+    :key="toolkit.id"
+    @click="openToolkitPDF(toolkit.pdfUrl, toolkit.name)"
+    class="relative bg-cross-lines border-2 border-[#0C0D0D] rounded-lg p-4 shadow-[4px_4px_0px_#0C0D0D] hover:shadow-[6px_6px_0px_#0C0D0D] transition-all duration-300 cursor-pointer focus:outline-none focus:ring-2 focus:ring-[#087BB4] focus:ring-offset-2"
+    :aria-label="`Download ${toolkit.name} accessibility toolkit PDF`"
+  >
+    <div class="game-resource-card-banner h-1/3" aria-hidden="true"></div>
+    <div class="flex items-center justify-between mb-3">
+      <h3 class="z-50 text-[18px] font-[600] text-white">
+        {{ toolkit.name }}
+      </h3>
+      <span aria-hidden="true" class="z-50">
+        <span v-if="toolkit.platformIcon == 'desktop'">
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width="24"
+            height="24"
+            fill="white"
+            class="bi bi-display"
+            viewBox="0 0 16 16"
+          >
+            <path
+              d="M0 4s0-2 2-2h12s2 0 2 2v6s0 2-2 2h-4q0 1 .25 1.5H11a.5.5 0 0 1 0 1H5a.5.5 0 0 1 0-1h.75Q6 13 6 12H2s-2 0-2-2zm1.398-.855a.76.76 0 0 0-.254.302A1.5 1.5 0 0 0 1 4.01V10c0 .325.078.502.145.602q.105.156.302.254a1.5 1.5 0 0 0 .538.143L2.01 11H14c.325 0 .502-.078.602-.145a.76.76 0 0 0 .254-.302 1.5 1.5 0 0 0 .143-.538L15 9.99V4c0-.325-.078-.502-.145-.602a.76.76 0 0 0-.302-.254A1.5 1.5 0 0 0 13.99 3H2c-.325 0-.502.078-.602.145"
+            />
+          </svg>
+        </span>
+        <span v-else-if="toolkit.platformIcon == 'controller'">
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width="24"
+            height="24"
+            fill="white"
+            class="bi bi-controller"
+            viewBox="0 0 16 16"
+          >
+            <path
+              d="M11.5 6.027a.5.5 0 1 1-1 0 .5.5 0 0 1 1 0m-1.5 1.5a.5.5 0 1 0 0-1 .5.5 0 0 0 0 1m2.5-.5a.5.5 0 1 1-1 0 .5.5 0 0 1 1 0m-1.5 1.5a.5.5 0 1 0 0-1 .5.5 0 0 0 0 1m-6.5-3h1v1h1v1h-1v1h-1v-1h-1v-1h1z"
+            />
+            <path
+              d="M3.051 3.26a.5.5 0 0 1 .354-.613l1.932-.518a.5.5 0 0 1 .62.39c.655-.079 1.35-.117 2.043-.117.72 0 1.443.041 2.12.126a.5.5 0 0 1 .622-.399l1.932.518a.5.5 0 0 1 .306.729q.211.136.373.297c.408.408.78 1.05 1.095 1.772.32.733.599 1.591.805 2.466s.34 1.78.364 2.606c.024.816-.059 1.602-.328 2.21a1.42 1.42 0 0 1-1.445.83c-.636-.067-1.115-.394-1.513-.773-.245-.232-.496-.526-.739-.808-.126-.148-.25-.292-.368-.423-.728-.804-1.597-1.527-3.224-1.527s-2.496.723-3.224 1.527c-.119.131-.242.275-.368.423-.243.282-.494.575-.739.808-.398.38-.877.706-1.513.773a1.42 1.42 0 0 1-1.445-.83c-.27-.608-.352-1.395-.329-2.21.024-.826.16-1.73.365-2.606.206-.875.486-1.733.805-2.466.315-.722.687-1.364 1.094-1.772a2.3 2.3 0 0 1 .433-.335l-.028-.079zm2.036.412c-.877.185-1.469.443-1.733.708-.276.276-.587.783-.885 1.465a14 14 0 0 0-.748 2.295 12.4 12.4 0 0 0-.339 2.406c-.022.755.062 1.368.243 1.776a.42.42 0 0 0 .426.24c.327-.034.61-.199.929-.502.212-.202.4-.423.615-.674.133-.156.276-.323.44-.504C4.861 9.969 5.978 9.027 8 9.027s3.139.942 3.965 1.855c.164.181.307.348.44.504.214.251.403.472.615.674.318.303.601.468.929.503a.42.42 0 0 0 .426-.241c.18-.408.265-1.02.243-1.776a12.4 12.4 0 0 0-.339-2.406 14 14 0 0 0-.748-2.295c-.298-.682-.61-1.19-.885-1.465-.264-.265-.856-.523-1.733-.708-.85-.179-1.877-.27-2.913-.27s-2.063.091-2.913.27"
+            />
+          </svg>
+        </span>
+        <span v-else-if="toolkit.platformIcon == 'mobile'">
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width="24"
+            height="24"
+            fill="white"
+            class="bi bi-phone"
+            viewBox="0 0 16 16"
+          >
+            <path
+              d="M11 1a1 1 0 0 1 1 1v12a1 1 0 0 1-1 1H5a1 1 0 0 1-1-1V2a1 1 0 0 1 1-1zM5 0a2 2 0 0 0-2 2v12a2 2 0 0 0 2 2h6a2 2 0 0 0 2-2V2a2 2 0 0 0-2-2z"
+            />
+            <path d="M8 14a1 1 0 1 0 0-2 1 1 0 0 0 0 2" />
+          </svg>
+        </span>
+      </span>
+    </div>
+    <p class="text-[14px] text-body py-3">
+      {{ toolkit.description }}
+    </p>
+    <div class="bg-[#E5F0F5] rounded px-2 py-1 inline-block">
+      <span class="text-[14px] text-[#087BB4] font-semibold">{{
+        toolkit.category
+      }}</span>
+    </div>
+  </button>
+</template>
